requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.1.0
boto3==1.35.23
botocore==1.35.23
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
cloudinary==1.39.1
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
distlib==0.3.7
distro==1.9.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==4.2.11
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-extensions==3.2.3
django-pwa==2.0.0
django-storages==1.14.4
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.2.2
filelock==3.13.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
lxml==5.3.1
MarkupSafe==2.1.3
mysqlclient==2.2.7
numpy==2.2.3
openai==1.30.3
packaging==24.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.2.0
platformdirs==3.11.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
s3transfer==0.10.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
virtualenv==20.24.6
whitenoise==6.6.0
